requirements.txt
attrs==22.1.0
autopep8==1.7.0
certifi==2022.9.24
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
identify==2.5.5
idna==3.4
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
logger==1.4
MarkupSafe==2.1.1
nodeenv==1.7.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
py==1.11.0
pycodestyle==2.9.1
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
toml==0.10.2
tomli==2.0.1
urllib3==1.26.12
virtualenv==20.16.5
Werkzeug==2.2.2
zipp==3.8.1
